+++ common/src/main/java/com/selfdot/cobblemontrainers/screen/Screen.java
@@ -26,16 +26,20 @@ public abstract class Screen {
         this.columns = columns;
         ScreenUtils.fill(inventory, Items.GRAY_STAINED_GLASS_PANE);
         for (int i = 0; i < columns; i++) {
-            inventory.setStack(i, new ItemStack(Items.BLACK_STAINED_GLASS_PANE));
-            inventory.setStack(((rows - 1) * columns) + i, new ItemStack(Items.BLACK_STAINED_GLASS_PANE));
+            inventory.setStack(i, ScreenUtils.withoutAdditional(Items.BLACK_STAINED_GLASS_PANE));
+            inventory.setStack(
+                ((rows - 1) * columns) + i, ScreenUtils.withoutAdditional(Items.BLACK_STAINED_GLASS_PANE)
+            );
         }
         for (int i = 1; i < rows - 1; i++) {
-            inventory.setStack(columns * i, new ItemStack(Items.BLACK_STAINED_GLASS_PANE));
-            inventory.setStack((columns * i) + (columns - 1), new ItemStack(Items.BLACK_STAINED_GLASS_PANE));
+            inventory.setStack(columns * i, ScreenUtils.withoutAdditional(Items.BLACK_STAINED_GLASS_PANE));
+            inventory.setStack(
+                (columns * i) + (columns - 1), ScreenUtils.withoutAdditional(Items.BLACK_STAINED_GLASS_PANE)
+            );
         }
         if (returnsTo != null) {
             backButtonSlot = ((rows - 2) * columns) + (columns / 2);
-            ItemStack itemStack = new ItemStack(Items.BARRIER);
+            ItemStack itemStack = ScreenUtils.withoutAdditional(Items.BARRIER);
             itemStack.setCustomName(Text.literal("Back"));
             inventory.setStack(backButtonSlot, itemStack);
         }
@@ -57,7 +61,7 @@ public abstract class Screen {
     }
 
     protected void setSlot(Inventory inventory, int index, Item item, String name) {
-        ItemStack itemStack = new ItemStack(item);
+        ItemStack itemStack = ScreenUtils.withoutAdditional(item);
         itemStack.setCustomName(Text.literal(name));
         inventory.setStack(index, itemStack);
     }